# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Revues API
  version: 0.1.0
  description: 'Contrat HTTP JSON de Revues (`/api/v1`). Source de vérité pour le codegen

    serveur (oapi-codegen) et le client TypeScript (WP-005).

    '
servers:
- url: /api/v1
  description: API versionnée
...
paths:
...
  /bootstrap:
    get:
      operationId: getBootstrap
      tags:
      - auth
      summary: Bootstrap session + CSRF
      description: 'Renvoie l''état d''auth courant et un `csrf_token` utilisable par le front.

        Sans session, pose un cookie guest court et dérive le CSRF de ce token.

        Avec session, renvoie l''utilisateur et le CSRF dérivé du cookie `revues_session`.

        '
      responses:
        '200':
          description: État bootstrap
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /me:
    get:
      operationId: getMe
      tags:
      - auth
      summary: Utilisateur authentifié courant
      description: 'Exige une session valide. Renvoie le profil + CSRF (même dérivation que bootstrap).

        '
      responses:
        '200':
          description: Utilisateur courant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/login:
    post:
      operationId: postAuthLogin
      tags:
      - auth
      summary: Connexion email + mot de passe
      description: 'Authentifie un compte local. CSRF requis (`X-CSRF-Token`, guest ou session).

        Pose le cookie `revues_session` en cas de succès.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Session créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/register:
    post:
      operationId: postAuthRegister
      tags:
      - auth
      summary: Inscription email + mot de passe
      description: 'Crée un compte local si l''email est autorisé (whitelist / bootstrap).

        CSRF guest requis. Pose `revues_session` en cas de succès.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Compte créé et session ouverte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout:
    post:
      operationId: postAuthLogout
      tags:
      - auth
      summary: Déconnexion
      description: 'Invalide la session courante et expire le cookie. CSRF session requis.

        '
      responses:
        '204':
          description: Déconnecté
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /orgs:
    get:
      operationId: listOrganizations
      tags:
      - orgs
      summary: Organisations de l'utilisateur
      description: 'Liste les organisations dont l''utilisateur est membre, l''org active

        de session (si définie), un défaut suggéré (`revues_last_org`), les

        invitations en attente, et si la création self-service est possible

        (aucune appartenance encore — parité pré-rewrite).

        '
      responses:
        '200':
          description: Liste org + invitations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createOrganization
      tags:
      - orgs
      summary: Créer une organisation (onboarding)
      description: 'Self-service : uniquement si l''utilisateur n''appartient encore à

        aucune organisation. Pose le créateur en `owner` et active l''org

        sur la session. CSRF requis.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organisation créée et activée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Non authentifié ou identifiants invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: CSRF manquant/invalide ou refus
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Conflit métier (ex. version publiée immuable)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Erreur interne
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorBody:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          example: validation_failed
        message:
          type: string
          example: Requête invalide.
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    User:
      type: object
      required:
      - id
      - login
      - email
      - display_name
      - role
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        avatar_url:
          type: string
        role:
          type: string
          description: Rôle global (admin|editor|reader)
    BootstrapResponse:
      type: object
      required:
      - authenticated
      - csrf_token
      - github_oauth_enabled
      properties:
        authenticated:
          type: boolean
        csrf_token:
          type: string
          description: Jeton CSRF à envoyer en header X-CSRF-Token sur les mutations
        user:
          $ref: '#/components/schemas/User'
        github_oauth_enabled:
          type: boolean
          description: true si REVUES_GITHUB_CLIENT_ID/SECRET sont configurés
        redirect:
          type: string
          description: Chemin SPA suggéré si déjà authentifié
    MeResponse:
      type: object
      required:
      - user
      - csrf_token
      properties:
        user:
          $ref: '#/components/schemas/User'
        csrf_token:
          type: string
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RegisterRequest:
      type: object
      required:
      - email
      - display_name
      - password
      - password_confirm
      properties:
        email:
          type: string
          format: email
        display_name:
          type: string
        password:
          type: string
          format: password
        password_confirm:
          type: string
          format: password
    AuthSuccessResponse:
      type: object
      required:
      - user
      - csrf_token
      - redirect
      properties:
        user:
          $ref: '#/components/schemas/User'
        csrf_token:
          type: string
        redirect:
          type: string
          description: Chemin SPA post-login (org onboarding ou accueil)
    Organization:
      type: object
      required:
      - id
      - name
      - slug
      - role
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        role:
          type: string
          description: Rôle org (owner|admin|member)
        ui_subject_label:
          type: string
        ui_run_label:
          type: string
    OrganizationInvitation:
      type: object
      required:
      - id
      - organization_id
      - organization_name
      - org_role
      - created_at
      properties:
        id:
          type: integer
          format: int64
        organization_id:
          type: integer
          format: int64
        organization_name:
          type: string
        org_role:
          type: string
        created_at:
          type: string
    OrganizationListResponse:
      type: object
      required:
      - organizations
      - invitations
      - can_create
      properties:
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        invitations:
          type: array
          items:
            $ref: '#/components/schemas/OrganizationInvitation'
        active_organization_id:
          type: integer
          format: int64
          nullable: true
          description: Org active sur la session (null si pending)
        default_organization_id:
          type: integer
          format: int64
          nullable: true
          description: Suggestion depuis cookie revues_last_org si membre
        can_create:
          type: boolean
          description: true si l'utilisateur n'a encore aucune appartenance
        redirect:
          type: string
          description: Chemin SPA suggéré selon le nombre d'orgs
    CreateOrganizationRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          description: Optionnel — dérivé du nom si omis
...
    OrganizationActionResponse:
      type: object
      required:
      - organization
      - redirect
      properties:
        organization:
          $ref: '#/components/schemas/Organization'
        redirect:
          type: string
          description: Chemin SPA après action (accueil)
...